Refactored configuration system.
