+++ chalice_restful/__init__.py
@@ -1,2 +1,2 @@
-from .core import Api, Resource, route
-from .configuration import authorized
+from .core import Api, Resource, route, config
+from .authorization import authorized
